requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==4.0
sqlparse==0.4.2
eth-abi==2.1.1
PyYAML==6.0
psycopg2==2.9.1
web3==5.20.0
djangorestframework==3.12.2
pytz==2021.3
drf-yasg==1.20.0
tronpy==0.2.6
gunicorn==20.1.0